Day 11 | Proved example part two

+++ implementation/src/main/kotlin/io/github/tomplum/aoc/simluator/monkey/Monkey.kt
@@ -1,8 +1,10 @@
 package io.github.tomplum.aoc.simluator.monkey
 
+import java.math.BigInteger
+
 data class Monkey(
     val id: Int,
-    val items: MutableList<Int>,
+    val items: MutableList<BigInteger>,
     val operation: MonkeyOperation,
     val test: MonkeyTest
 )
