+++ implementation/src/test/kotlin/io/github/tomplum/aoc/simluator/monkey/MonkeyOperationTest.kt
@@ -4,40 +4,41 @@ import assertk.assertThat
 import assertk.assertions.isEqualTo
 import org.junit.jupiter.api.Test
 import org.junit.jupiter.api.assertThrows
+import java.math.BigInteger
 
 class MonkeyOperationTest {
     @Test
     fun multiplyByInteger() {
         val operation = MonkeyOperation("  Operation: new = old * 19")
-        val newWorryLevel = operation.execute(2)
+        val newWorryLevel = operation.execute(BigInteger.valueOf(2))
         assertThat(newWorryLevel).isEqualTo(38)
     }
 
     @Test
     fun addToInteger() {
         val operation = MonkeyOperation("  Operation: new = old + 6")
-        val newWorryLevel = operation.execute(26)
+        val newWorryLevel = operation.execute(BigInteger.valueOf(26))
         assertThat(newWorryLevel).isEqualTo(32)
     }
 
     @Test
     fun multiplyByOld() {
         val operation = MonkeyOperation("  Operation: new = old * old")
-        val newWorryLevel = operation.execute(10)
+        val newWorryLevel = operation.execute(BigInteger.valueOf(10))
         assertThat(newWorryLevel).isEqualTo(100)
     }
 
     @Test
     fun addToOld() {
         val operation = MonkeyOperation("  Operation: new = old + old")
-        val newWorryLevel = operation.execute(10)
+        val newWorryLevel = operation.execute(BigInteger.valueOf(10))
         assertThat(newWorryLevel).isEqualTo(20)
     }
 
     @Test
     fun unknownOperator() {
         val operation = MonkeyOperation("  Operation: new = old - 5")
-        val e = assertThrows<IllegalArgumentException> {  operation.execute(10) }
+        val e = assertThrows<IllegalArgumentException> {  operation.execute(BigInteger.valueOf(10)) }
         assertThat(e.message).isEqualTo("Unknown monkey operation operator value [-]")
     }
 }
